test: concentrate phase 5 cli owners
Split stale mixed CLI owners into focused command, click-app, formatting, helper, and query-mode integration files; trim embed/run overlap; and move non-CLI regressions to their proper owners.

Verification:
- nix develop -c ruff check .
- nix develop -c pytest -q -n 0 tests/unit/cli tests/unit/ui tests/unit/rendering tests/unit/showcase tests/integration/test_site.py tests/integration/test_site_laws.py tests/integration/test_cli_query_mode.py tests/infra/test_storage_records.py tests/unit/core/test_lazy_exports.py tests/unit/pipeline/test_pipeline_ids.py
- nix develop -c pytest -q -n 0
